requirements.txt
pygame-ce
zengl
mili-ui
